+++ example/lib/screens/main/pages/settingsAndAbout/settings_and_about.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../../../shared/components/loading_indicator.dart';
@@ -11,6 +13,7 @@ class SettingsAndAboutPage extends StatefulWidget {
 }
 
 class _SettingsAndAboutPageState extends State<SettingsAndAboutPage> {
+  final creditsScrollController = ScrollController();
   final Map<String, String> _settings = {
     'Reset FMTC On Every Startup\nDefaults to disabled': 'reset',
     "Bypass Download Threads Limitation\nBy default, only 2 simultaneous bulk download threads can be used in the example application\nEnabling this increases the number to 10, which is only to be used in compliance with the tile server's TOS":
@@ -65,19 +68,69 @@ class _SettingsAndAboutPageState extends State<SettingsAndAboutPage> {
                         ),
                 ),
                 const SizedBox(height: 24),
-                const Header(
-                  title: 'App Information & Credits',
+                Row(
+                  children: [
+                    const Header(
+                      title: 'App Credits',
+                    ),
+                    const SizedBox(width: 12),
+                    TextButton.icon(
+                      onPressed: () {
+                        showLicensePage(
+                          context: context,
+                          applicationName: 'FMTC Demo',
+                          applicationVersion:
+                              'for v5.1.0\n(on ${Platform().operatingSystemFormatted})',
+                          applicationIcon: Image.asset(
+                            'assets/icons/ProjectIcon.png',
+                            height: 48,
+                          ),
+                        );
+                      },
+                      icon: const Icon(Icons.info),
+                      label: const Text('Show Licenses'),
+                    ),
+                  ],
                 ),
                 const SizedBox(height: 12),
-                const Text(
-                  "An example application for the 'flutter_map_tile_caching' project, built by Luka S (JaffaKetchup).\n",
-                ),
-                const Text(
-                  'Many thanks go to all my donors, including:\n - @ibrierley\n - @tonyshkurenko\nIf you want to support me, any amount is appriciated! Please visit the GitHub repository for donation/sponsorship options.\n\nThanks to all of the dependencies used in both this app and the main library. In addition, thanks to Nominatim: their services are used to retrieve the location of a recovered download on the Recover screen.',
+                Expanded(
+                  child: SingleChildScrollView(
+                    controller: creditsScrollController,
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: const [
+                        Text(
+                          "An example application for the 'flutter_map_tile_caching' project, built by Luka S (JaffaKetchup). Tap on the above button to show more detailed information.\n",
+                        ),
+                        Text(
+                          "Many thanks go to all my donors, whom can be found on the documentation website. If you want to support me, any amount is appriciated! Please visit the GitHub repository for donation/sponsorship options.\n\nYou can see all the dependenices used in this application by tapping the 'Show Licenses' button above. In addition to the packages listed there, thanks also go to:\n - Nominatim: their services are used to retrieve the location of a recoverable download on the 'Recover' screen\n - OpenStreetMap: their tiles are the default throughout the application\n - Inno Setup: their software provides the installer for the Windows version of this application",
+                        ),
+                      ],
+                    ),
+                  ),
                 ),
               ],
             ),
           ),
         ),
       );
+}
+
+extension on Platform {
+  String get operatingSystemFormatted {
+    switch (Platform.operatingSystem) {
+      case 'android':
+        return 'Android';
+      case 'ios':
+        return 'iOS';
+      case 'linux':
+        return 'Linux';
+      case 'macos':
+        return 'MacOS';
+      case 'windows':
+        return 'Windows';
+      default:
+        return 'Unknown Operating System';
+    }
+  }
 }