+++ lib/src/root/import.dart
@@ -23,7 +23,7 @@ class RootImport {
   ///
   /// Uses the platform specifc file picker. Where supported, limits file extension to [fileExtension] ('fmtc' by default), otherwise any file can be selected as a fallback.
   ///
-  /// It is recommended to leave [emptyCacheBeforePicking] as the default (`true`). Otherwise, the picker may use cached files as opposed to the real files, which may yield unexpected results. This is only effective on Android and iOS - other platforms cannot use cache.
+  /// It is recommended to leave [emptyCacheBeforePicking] as the default (`true`). Otherwise, the picker may use cached files as opposed to the real files, which may yield unexpected results. This is only effective on Android and iOS - other platforms cannot use caching.
   ///
   /// If any files are selected, a [Map] is returned: where the keys are the the selected filenames (without extensions), and the values will resolve to a [bool] specifying whether the import was successful or unsuccessful. Otherwise `null` will be returned.
   Future<Map<String, Future<bool>>?> withGUI({
@@ -67,22 +67,23 @@ class RootImport {
   ///
   /// The output specifies whether the import was successful or unsuccessful.
   Future<bool> manual(File inputFile) async {
-    Future<bool> returnError(StoreManagement storeManagement) async {
+    Future<bool> error(StoreManagement storeManagement) async {
       await storeManagement.deleteAsync();
       return false;
     }
 
-    final String path = inputFile.absolute.path;
-    final String storeName = p.basenameWithoutExtension(path);
+    final String storeName =
+        p.basenameWithoutExtension(inputFile.absolute.path);
     final StoreManagement storeManagement =
         StoreDirectory(_rootDirectory, storeName, autoCreate: false).manage;
 
     if (!await compute(_import, {
-      _rootDirectory.access.stores > storeName: await File(path).readAsBytes(),
-    })) return returnError(storeManagement);
+      _rootDirectory.access.stores > storeName:
+          await File(inputFile.absolute.path).readAsBytes(),
+    })) return error(storeManagement);
 
     if (await storeManagement.readyAsync) return true;
-    return returnError(storeManagement);
+    return error(storeManagement);
   }
 }